Adds point discarding and fixes alignment application bug
Sets code to now remove the worst 10% of matches on each iteration of
icp. In addition, previously in the main icp loop, the updated
alignments were being applied to the original set of points and not with
the updated matches (registration.cpp in 'icp' function), this has now
been fixed.